ros-pcl: plugin to exchange points clouds with ROS
This is currently a basic version. It will check the available point
clouds only once on startup and publish them then as they come. This
most likely needs some sanity timer to publish only every n-th update.
At the moment it supports distribution only from Fawkes to ROS, not the
other way around.

The plugin was tricky and requires some compilation tricks because of
two things:
1. PointCloud is a template class which does not inherit from a common
base class. Hence storing them in a single data structure requires an
adapter.
2. The standalone PCL does not mix well with ROS because both contain
incompatible version of std_msgs and sensor_msgs type. The trick to get
this to work is to confine the two types of types in separate modules
and tweak the include order to be just right...