tabletop-objects: detect objects on a table surface
This is an intial implementation of a tabletop object segmentation,
detection, and maybe eventually recognition. Atm we detect clusters on a
table surface applying intensive pre-filtering. The implementation has
not been optimized and result data is not yet written to the blackboard.
It can be visualized by looking at rviz using the ros-pcl plugin.