tabletop-objects: avoid deletion of input pointcloud on finalize
The Fawkes RefPtr for the point cloud provided as input by the PCL
manager must be converted into a Boost shared_ptr to serve as input for
PCL. When resetting the shared_ptr, the cloud would be freed causing
segfaults if other processes still accessed the cloud (as was the case
since openni-pointcloud still wrote to it). Provide a little utility
method that creates a non-deleting shared_ptr to avoid this problem.