tabletop-objects: unify variable naming
Instead of using __this_style convert to using this_style_. It is used
throughout PCL and I gave it a try throughout the PCL code and actually
like it. It is less likely to conflict with system defines and it's
still as easy (some might say even easier) to read.