tabletop-objects: check angle between table normal an Z axis
Rather than only verifying that the Z component of table normal is
dominant apply a criterion on the angle between the table normal and the
Z axis in the base_link frame. This allows for a much stricter check of
the parallelism of the table to the ground plane. Currently restrict to
0.1 rad (based on my tripod Kinect). Need to check on robot if we can
decrease this further.