tabletop-objects: ICP table model fitting
Use ICP to fit an on-the-fly generated table model. After a few
optimizations (heavy downsampling, ICP parameters, getting a good
initial guess of the transformation, passing model and detected cloud in
the proper order...) finally speedy enough, but results are
unsatisfactory. It is easily tricked if the camera frustrum cuts off the
table at the lower end and therefore creates a virtual straight line and
at the same time the camera or robot is slightly turned, so that the
table is not actually parallel to the robot but in an angle. Also, it is
tricked for example if there are two tables alongside each other with
the long edge. The result is a large convex hull and table plane (point
cloud) causing the model to shift more towards the overall center.