tabletop-objects: replace ICP with polygon method
Removed ICP, performance wasn't good, especially table orientation was
bad, which is one of the more crucial parameters. Implemented a first
version that focuses on processing the convex hull of the table to
determine the alignment edge of the table. Then fit the table model to
that edge. Currently only visualization, will add passing of the
determined table parameters soon.