utils: add TimeTracker::ping_abort() method
If something happens in a timing step that you rather don't want to be
included in your time tracking result (e.g. premature abort of a
procedure due to an error) then this new method can help. When called
it'll end the current class measuring period without recording the
result.