utils: auto-master capability for SharedMemoryRegistry
Instead of having to insist on either being master or now, just assume
that the shared memory segment should be created if it does not exist,
and to re-use it if it does. The whole segment is protected by a lock
semaphore. Now the same registry can be properly used by multiple Fawkes
instances, as is desired for example when separating an (image) buffer
generating and processing thread in two processes.