openni-image: write stable buffer, move to prepare hook
OpenNI uses multiple buffers internally. If the meta data (from which we
get the pointer to the data buffer) is not updated in every loop we use
the wrong buffer every Nth frame. This means that the buffer could
contain half fresh and half old data and did not match the point cloud.