openni-data: merge image and point cloud plugins
In order to provide XYZRGB point cloud buffers close interaction of
image and depth buffers are necessary showing the inherent dependencies
of the two. The most simple solution out of all the possibilities is to
merge data providing plugins, which this commit does. openni-image and
openni-pointcloud are now deprecated. The new openni-data plugin will
take care of generating only the data required.

The new plugin now also provides XYZRGB buffers and point clouds.

Note: this commit also changes the buffer IDs and depending plugins need
to be adapted.