ros-pcl: keep point cloud open while used
The plugin only quickly glanced at the buffer each time if an update was
required. This breaks the usage detection of the openni-data plugin
because the refptr use count is not permanently increased. To fix store
the storage adapter to have a permanent pointer to the point cloud.
Close it if it is no longer needed so that data is not generated in that
case.