aspect: improved point cloud type checking for old compilers
The PointCloudManager relied on dynamic_cast to check the point cloud data
type. On older compilers this sometimes fails for no good reason. The most
likely explanation is that the same generated type (instantiated from a
template) gets two type_info instances, each with the same information, but
still separate instances, hence making dynamic_cast fail. Now we compare
the typeid string and do a static_cast if it matches.