baseapp: catch excpetions in individual wakeup hooks
The main thread would encapsulate wakeup and wait calls for all hooks in a
single try-catch block. This will cause later hooks not to be called if an
error occurs in an earlier hook, for example if a thread is marked as bad
because it runs for too long. By catching the exceptions per hook we can
continue the current loop as good as we can without the rogue thread(s).