core: fix "running with another barrier" bug and segfaults
Sometimes recovering a bad thread could cause a then-permanent error that some
thread could no longer be woken up because it was supposedly running already
with another barrier (even though it actually didn't). The problem was caused
by a combination of problems. Briefly, the problem is that the last thread
reaching the InterruptibleBarrier would wake all threads waking at the
barrier. However, if the main thread would be woken up before another thread
(not unlikely, especially on a single-core machine), the thread would not have
a chance to be woken up in time. Now, the formerly bad thread (which must have
waited for the same barrier in the same hook) recovers, causing the
"bad barrier" to be deleted, now leaving the symptomatic thread waiting for a
no longer existing wait condition. In the "best" case the program would
continue just without this one thread. In other situations, especially on our
Robotino with an older Linux version, this would lead to segfauls on a
following malloc call.

The resolution consist of two parts. First the interruptible barrier
synchronizes on a regular barrier after all threads have completed and there
was no timeout. Additionally the thread list's exception handling and ordering
of unsetting the barrier was modified.