openrave: simplify OpenRAVE detection
Use only pkg-config to determine cflags and ldflags for OpenRAVE and
Python. Mixing with the config-files is risky, they might not exist on
other installations or not be in the path. Also, the OpenRAVE tools gave
invalid flags for x86_64 (included 32bit lib dir).

Now only a minimum version needs to be specified in a simpler way. Also,
version suffixes are now detected by explicitly searching for
openrave-core in the list of available packages.