amcl: properly initialize AMCLLaser
There were two instances of AMCLLaser, the first which was properly
initialized with the model parameters and the second which was actually
used later -- without the model parameters! Fix this by removing the
second instance altogether. It was a remnant from supporting more than
one laser in the ROS amcl node.