robotino: fix odometry reset and do reset on startup
The odometry reset was resetting the motor encoders to zero instead of
zeroing the odometry. In addition to fix this we now reset the odometry
once on startup. It is without risk at that point in time and makes
things easier later on when at least starting with a nice odometry
system.