amcl: add grace period to still call updates after stopping to move
There are distance and angle thresholds on the robots movement and amcl
updates are only performed if the robot's motion exceeds these
thresholds. We now add a third criterion based on time. After the robot
has stopped to move, amcl updates will still be performed for the given
time. This is useful especially on robots with mediocre odometry, where
continued laser data updates are required after a motion to catch up
with drift errors.