amcl: remove assertion causing trouble on robotino
The Robotino compiler is broken and hence exception code is buggy and the
assertion exception cannot be caught. To remedy this the assertion has been
commented out for now and pz is set to zero. However, we are experiencing
problems from time to time with AMCL, and it might be this very assertion
that could lead to its cause. Hence consider this commit to be preliminary
until further investigation.