libconfig: only pass one file path to load
Configuration loading now only takes a single file name. If another one
is required it must either be derived from the given filename,
configured in the given file, or otherwise be determined, for example
taking the host's name. This allows us to get rid of some cruft never
used and in the future no longer supported.