libconfig: host-specific configuration implemented
The host-specific configuration file is a specifically designated file
which is now read separate from the existing configuration, and it is
read last allowing it to overwrite any other config values. There can
only be a single host-specific configuration file and it may not include
any other files.

If changes are made to the configuration using methods like set_float()
they modify the host-specific configuration and are written to the
host-specific configuration file. Erasing likewise erases values in the
host-specific config file.

Setting or erasing default values is not supported by the YAML
configuration implementation.