libconfig: type confirmation based on already determined type
Use the already determined type to decide on the type. Checking if a
value is merely convertible to the desired type makes the result of
calling the different methods depending on the order, e.g. for the
numerical value 0 is_int(), is_float(), and is_uint() would all return
true, while we only want it to be true for is_uint().