katana: add correct check-for-final for gripper
Cannot check gripper like other motors, i.e. checking if
target position is reached (which btw is incorrect for "closed"
in the kni-configfile, should be +13xxx instead of -16xxx). Gripper
should be and now is also considered final, if motor has not moved
for a while. This behaviour was implemented before, but not taken
over when introducing the Controller-class.
TODO: Check movement over time in future versions. Currently we check
if movement stopped for 3 "cycles", just the way it was originally
implemented.