openrave: fix openrave detection
Various bug in the detection were fixed. Check for OPENRAVE_MIN_VERSION
allways failed due to a missing "$(if ...", therefore going into check
for different naming of openrave-corei everytime. That never checked the
detected version, whether it was empty (i.e. returning "openrave-core")
or any number (which could be ANY arbitrary version), thus basically
just returning "true" if "pkg-config --list-all | grep openrave-core"
returned non-empty.

The OpenRAVE version is an infix, not a suffix, as it is part of
the program name. Thus, we would need to grep with a regular expression
to find versioned occurences of 'openrave-core', making it too
complicated and yet not sufficient, e.g. for MIN_VERSION=0.6.4
we would need REGEX='[0-9]\{1}.[6-9]\{1}' and grep with
"grep 'openrave$(REGEX)-core" to get at least 0.6,  but this
would fail for version 1.1 (not out yet, but still valid point).
The effort would not be worth it. MIN_VERSION implicits a
specific program version, i.e. 0.6.4 requires openrave0.6, and
0.7.0 requires openrave0.7. We now check version with an infix
gained from a substring of MIN_VERSION.