openrave: extend caesar model by cover-plate to reduce collisions
Not the whole cover-plate (the one with the holes for katana and
wifi-antennas), but only the front of it, which is sufficient and
serves the same purpose: Avoid collision with 2nd link of katana
with that part of the cover.
This should also reduce the chances to hit a table with intermediate
arm links.