amcl: move map and config reading to new lib, crucial bug fix
Refactored reading of map and config values in a new utility library.
Along the way fixed a crucial bug. In the map reading code, as well as
in the visualization message creation code the meaning for free and
unknown cells were swapped. The wrong value of course made the map
utility library failed on the map. For example, the Bresenham raytracer
would always return zero as beam length to the next obstacle. We even
managed to continue this bug by modifying the uniform pose generator to
work on the wrong map data. Wow, what a debugging success this must have
been (on the pose generator) *sigh*