amcl: fix crucial initialization problem
There were members named laser_max_range and laser_max_range_. The
config value was written to the former but the latter was used. Probably
due to optimizer flags the latter was most of the time initialized to a
value just slightly unequal to zero. Hence the default max range did not
kick in, but still all values were being rejected because they were
greater than the allowed max range (about but not quite zero).
Convergence behavior is now similar to the AMCL ROS node.