blackboard: remove notifier wait conditions
If the event queue is being processed and actions during this processing
lead to new events (of the same type), a deadlock would occur. We relax
the assumption such that events have not necessarily been processed
after the event causing method returns, rather they might be queued.
This avoids the deadlock by making the wait conditions unnecessary.