skill: new katana_approach: approaching some object
This skill just approaches "forward" (+x) or "upward" (z). All params
are optional, but at least the param "dist" is recommende to set the
desired approach distance.
On failure, numerous other attempts are made with little derivations
from the original destination to still try an approach on a straight
line. All the params that are used to calculate the derivations can
be modified, see skill documentation for further information.

This is very useful, e.g. for lifting a grabbed object. Using this
skill one has a significantly higher probability to lift in a straight
line, which may even go a little forth/back. It also slims down
other complex skills, that called multiple approaches for lifting or
approaching, which now need only 1 call to this skill.