katana: fix final check for simultanious movement of joints
The current version had a more or less major bug. We take the
joint configurations of each point in a trajectory, then set
each joint individually. Setting single joints was treated in
the controller in a way that it regarded only this joint as active,
which would lead to only checking the "last set" joint to be checked
for final position. So sometimes a "final" was returned true before
actually being finished.
This was not discovered because most times all joints finished at
the same time (due to RRT planning). In cases where it finished
prematurely, the consequences were not significant as for each new
plan the current values were fetched again. The only downside it had,
was that the openrave model was not allways accurate, as it gets the
last update right after movement-finish, leading to a position in
the Gui which did not represent the last position of the real Katana.