katana: goto_openrave_thread allows editing of plannerparams
It still needs to be implemented in an interface and act_thread,
such that users can set it via skills.
Setting plannerparams to "default" will take the default params.
It is intentionally not an empty string, as we also want to allow
empty strings as plannerparams, as in 'do not use additional params'.