utils: generic TopologicalMapGraph with RCSoft graph loader
On the way to a more generic topological map graph create a generic
TopologicalMapGraph class which is based on the existing RCSoftMapGraph
sans the RCSoft graph specific details. We also remove some old cruft
barely used (aliases) and introduce more powerful properties (a proper
mapping from key to value). We also now have an explicit edge
representation. For one it makes usage clearer and inconsistencies are
easier to spot. Edges now can also have properties (for example we can
have a travel speed for an edge, and a separate approach speed for
nodes).

The RCSoftMapGraph is no more. The parsing part is now in a loader
method which reads RCSoft graphs and represents them as a
TopologicalMapGraph instance.