utils: support topological graphs specified using YAML
The RCSoft map graph specification showed the path but has some
drawbacks. First the choice of XML makes it hard to read or modify the
graphs as text manually. Second the implicit edge definition by only
specifying child nodes for a node is harder to read (it is not easy to
spot which connections actually exist) or check for inconsistencies.

The new YAML format (to be documented in the wiki soon) allows for a
much shorter specification which is more readable. For example, for the
KBSG graph it requires less than half the number of lines. It also
improves on properties, where it now stores key/value pairs (but
singular properties like "GrabNode" are still possible). The new format
also explicitly supports directed edges.