tf: if ROS bullet is available use it, fixes ticket #151
ROS and regular bullet versions are incompatible. Even mixing different
modules linked against different versions already causes problems. The
only solution is to decide for either one. The main tf config file will
now check if ROS' bullet is available, and if it is use that version. It
is required for ROS integration and in particular ros-tf.