utils: assert that topological map graph is fully connected
Make sure that all nodes are reachable from everywhere. Today I had
tricked myself by having two connected nodes (A,B) and another two
(C,D). Since D was lying on the edge from A to B it seemed that the
graph was connected, while in fact they were not. The added check now
verifies that the graph is fully connected. The implemented method is
not particularly efficient, but will work for reasonable graph sizes.