navgraph: add configurable post-tolerance reached travel time
After the final target node has been reached, i.e. the robot's position
is within the tolerance perimeter of the node, we have a configurable
time we let the colli/navi/local path planner continue moving to get
closer to the point. We need to abort after some time, otherwise
tolerance would be void, but we also should let it go for a little more
time, otherwise we would almost always have an error of about the same
magnitude as the tolerance.