navgraph: port simple plan optimization
After generating a plan, the robot first needs to travel to the first
actual node from a free position within the environment. It can happen,
that this closest node lies in the opposite direction of the second
node, hence the robot needs to "go back" first, and only then starts
following the path. We can optimize this by removing the first node, so
that the robot directly travels to the second node which "lies on the
way".

Note that only the start node optimization has been ported, and only for
the very first node. There are situations were second and end node
optimizations could do more harm than good and it would forfeit the
expertise of the graph creator, hence leave this out. This can easily be
accomplished by adding more nodes and connections when necessary.