libconfig: raise critical error if yaml-cpp is missing
Similar to SQLite we really need YAML from now on. Therefore we raise a
critical error if yaml-cpp is not found as to make clear where the
problem comes from.

Recently, my FreeBSD port for yaml-cpp was accepted so we can still
build on FreeBSD out-of-the-box.
http://www.freebsd.org/cgi/query-pr.cgi?pr=ports/172146

Linux distros typically bring yaml-cpp, checked on Fedora and Ubuntu.
Since ROS uses this extensively and it is an external OS dependency
since Fuerte any platform supporting robots is likely to also bring this
library in the future.