skill_jumpstate: allow same syntax for subskills as for luaagent
In 'define_states' one can now define the skills-table the same
way as in luaagent, i.e. using skill-name instead of skill:
skills={{"subskill"}}
This is just to allow a homogenous syntax among the state-machines.

REMARK: this is done with help of skillenv, having knowledge about
all other skills. I do not know if this behaviour can be tolerated
or if skill_jumpstates should be a rather "clean" module for itself.
The latter would require more modifications to provide this funcionality,
maybe even using skill-queues like luaagents do.