baseapp: user-dependent shared memory registry
The shared memory registry's name is now determined based on the current
processes user name (with the fallback to the current behavior if this
cannot be determined, which should actually not happen). This allows
multiple instances of different users to run on the same hosts side by
side (to be tested by lab group).