config: add transforms example config for Kinect on tripod
If the kinect is mounted on a tripod the transforms towards base_link
must be statically setup. The config can be used to do this. It mimics
the transform chain we have on caesar, in particular having a pan/tilt
unit. It also makes it easy to adjust. For the height adjust the PTU
base height, then pan and tilt can be set individually.