libconfig: support for config value lists
To extend our capabilities towards what is possible with YAML the API
has been extended to also support lists of values for a single path.
Currently this is only implemented for the YAML config, the network
and SQLite configuration implementations will throw
NotImplementedExceptions for the related functions. Transmission of the
lists over the network is not yet supported.