mongodb-log: include/exclude parameters for point clouds
Unified the configuration names to use a dash, not an underscore. Added
include/exclude parameters that allows to be more specific about the
point clouds to record to the database. E.g. avoid storing XYZ and
XYZRGB point clouds from the same sensor, the latter is enough to have
all the data, the former can be used to reduce the amount of data, both
are just overhead, no benefit.