pcl-db-merge: merge and align point clouds from database
This new plugins reads in a specified number of point clouds, transforms
them to a common reference frame, and then aligns them more closely by
using PCL's ICP registration and some optimizing and biasing magic. The
resulting merged point cloud is again published for further processing,
e.g. by the tabletop-objects plugin.

In this intermediate state times for a particular run with the PR2 are
hardcoded, as well as all of the parameters. I can provide the database
if you want to test it. The parameterization is going to change soonish.