tabletop-objects: proper output in case of non-modeled detection
If no table model parameters can be estimated, i.e. if the table is out
of the frustrum of the envisioned sensor, for example because we are
working on reconstructed data, properly recomput the table centroid
based on the table cluster detected. Also fix publishing of the cluster
point cloud, it was not using the latest decided table points but rather
the intial guess.