tabletop-objects: consider viewpoint relation to table
For the extraction of clusters "above" the table we must determine the
relation of the viewpoint, i.e. the origin of the input point cloud's
reference frame, to the table centroid. The decision is based on the
relation in the Z axis direction in the base link frame, assuming this
frame is bound to the ground support plane of the scene (which for
domestic service robots holds). Then we refine our normal vector
orientation expectation and the filter based on this criterion.