ros: support asynchronous ROS main loop spinning
The spinner for the ROS main loop can now be asynchronous threads. While
we favor main loop integration, this has the problem that it blocks the
ROSNodeThread in the case of long-running service calls (for example if
a Fawkes plugin provides such a service). The new config file sets this
as the default. To be monitored.