mongodb: pass MongoDB connection manager to threads
Not only create a single MongoDB client connection for plugins with the
MongoDBAspect, but also provide acess to the MongoDB connection manager
to acquire more connections. Since a MongoDB client is not thread-safe
more clients must be created if a plugin uses multiple threads to write
to the database at the same time.