mongodb-log: add specialized and optimized transform logger
Logging of transforms over the blackboard can incur a considerable
overhead. On the PR2 it means doing ~4000 inserts/sec. This can be
vastly improved by storing all transforms in intervals of, for example,
five seconds. This reduces it to 100 inserts/5 sec. This is achieved by
using the new tf extensions to access the internal storage. It is copied
(and this is the only time in which the tf structures are locked) and
then written out. All transforms for one specific child frame in the
given time range are then stored in a single document.