clips-agent: generic skill calling support
A new utility function and fact templates allow to interact with the
skiller. Arbitrary skills can be called and facts are asserted to
react to state changes.

This again includes skill simulation support, meaning that all called
skills are assumed to be successful. The time each skill should take can
be configured.