libconfig: MemoryConfiguration and use in NetworkConfiguration
Added a new class that reuses YamlConfiguration data structures to
create an in-memory configuration store. This is now used in the
NetworkConfiguration as mirror config. Therefore we can now drop the
SQLiteConfiguration if and when we wish. It will also pave the way for
transmitting list values over the network.