navgraph: introduce final_tolerance property
When navigating via waypoints, we generally want a low tolerance on the
final node, but a high tolerance on all the waypoints. Thus, a new
config attribute and navgraph property called "final_tolerance" is
introduced. The regular tolerance value will be used for all nodes
except the last one, where final_tolerance will be used instead.