lua: graceful handling of code errors
If FSM code was broken, the thrown error was escalated to the top level.
This could leave a state machine in a non-reset state triggering
unforeseen behavior in the next run. Therefore, wrap invocations of user
implemented functions in xpcall to gracefully handle it and recognize it
as internal error. The error message still contains the stack trace with
this solution. Thanks for Victor Matare for reporting this error.