laser-cluster: make line removal optional
Line removal can now be disabled in the configuration. Often it is
simply not necessary, discriminating clusters is sufficient.
Along the way make the minimum number of inliers for line segments an
absolute value. For the small number of points this seems to work better
(is easier to configure) than the previous quota.