buildsys: detect if PCL was build with OpenMP support
Query linker for default library paths and search for libpcl_common.
Then check if this library was linked against the system's OpenMP
library. If it is, plugins and binaries linking against PCL might need
to add OpenMP flags if they use Eigen code, because in PCL it might use
OpenMP, but in the plugin/binary build not. This causes problems as
Eigen is a header-only library.