laser-pointclouds: set frame ID in every loop
There can be a race condition if the input (laser) and laser-pointclouds
plugin are loaded at the same time. Then, the frame_id has not
necessarily been set before init() is called and the frame ID remains
invalid. Setting it in each loop is a minor overhead and fixes this
problem.