bumblebee2: driver plugin for Bumblebee2 stereo camera
This plugin supports stereo processing by means of Triclops or OpenCV
(the former crashes on recent Linux system, a bug has been filed with
PTGrey but it does not seem there will be a quick fix).
The OpenCV stereo processor still uses the Triclops rectification, so
the Triclops SDK is required in any case. This is done to avoid time
consuming stereo calibration and just go with the (arcane) data that
comes with the camera.