bumblebee2: publish transforms for lenses
Given a base link publish transforms according to the camera specs for
the left and right lenses. One of the two is required to situate the
point clouds (and which depends on the stereo matcher used), the other
might still be used for vision processing and therefore we publish it.