clips: move environment maintenance to CLIPSEnvManager
By separatig the environment maintenance to a separate entity we can
more easily re-use it in the future.

This comes with an important additional change: environments are now
always for exclusive use and created on initial request. Shared
environments can now longer be created (they resulted in too many
problems and segfaults). A future extension will bring back the one use
case of shared environments: providing features to an environment.