clips-protobuf: use new CLIPSFeatureAspect to provide protobuf
The protobuf capabilities are now provided as a feature, i.e. a library
which can be requested from a CLIPS environment which is instantiated
through the CLIPSAspect.

This fixes in particular the problems with the proper
initialization/finalization order and associated segfaults if a user got
it wrong. The framework can now make sure that only proper sequences are
issued and can give helpful messages in case of a problem.