clips: add config feature (moving more code from clips-agent)
The clips plugin now provides a new basic config feature to access the
Fawkes configuration. It is based on code formerly residing in the
clips-agent plugin. The clips-agent has been adapted to the code move.
The config feature is implicitly loaded by the agent as it is vital to
its initialization.