colli: transfered MotorControl class from RCSoftX
This is basically a wrapper for controlling the motor. It is a "gateway"
to send commands, and get current and desired motor values (which requires
2 MotorInterfaces at the moment).
In the futuer, this could be skipped and the MotorInterfaces could be directly
accessed instead. But I want to keep it as close to RCSoftX as possible.