colli: add reading interface for desired motor values
The MotorInterface just provides current data, not desired position or
velocities. Those are needed in colli though.
Therefore we require 2 MotorInterfaces for now. Whether the motor plugin
should provide both (e.g. "Motor" and "Motor desired"), or the colli
plugin should make a writing "Motor desired" just for itself, or if
the MotorInterface shuold be extended to include all that information...
...remains to be seen :)