colli: make m_pColliTargetObj a reading, not writing interface
It is a 'client' in RCSoftX, so the equivalent would be a reading interface
in fawkes. However, the colli gets its target information from this interface!
In RCSoftX it was libmonaco that filled it, ordered by fawkes_sync.
We need to provide this here, too. I suggest a second thread that does that,
just to have it separated as it was in RCSoftX. It is definitely not the best
solution (could probably be integrated into colli_thread with less code), but
that is not the priority at the moment. I want the same exact thing as in
RCSoftX right now, with optimizations in a later step :)

+ Using the is_orient_at_target field of NavigatorInterface now to decide
about that behaviour.