colli: rename writing navi-interface to "colli data"
In RCSoftX it was used to just provide some minimal information about
the colli. It was not really the main Navigator Interface that accepted
commands and provided all the information to the outside.

It might be replaced at later point to be the actual "Navigator" interface,
but for the moment I'll stick closer to RCSoftX implementation. Maybe
I don't even need an interface for this..