MotorInterface: add fieds for desired vx,vx,omega
In addition to current motor information, it is sometimes useful to know
what the desired values are, like those that are received by TransRotMessages.
Makes sense, as the motor has acceleration and deceleration phases.
Some plugins already differenciate between current and desired motor values
(colli plugin does that, currently using 2 interfaces for that)