colli: parameterize pre_position_distance
If the target is not close (i.e. it is >=min_drive_rot_distance), we
don't go straight to the target but approach a pre-position instead.
The pre-position considers the target position and orientation, and moves
one "step" back. The distance of that "step" can now be adjusted in the
config.