cedar: add ROS node and topic connection monitoring
Given a model monitor if nodes are alive and connections for topic
communications have been established. Merge results into a ROS system
condition.

This is still missing a few features from my earlier SRI CEDAR, but it
is implemented in a much cleaner way, with better interfaces, and more
handling of things within CLIPS (rather than in C++ preprocessing).