BE hsm.lua: introduce cond,precond,cond_and_precond as suggested in #275
The documentation on "add_transitions" is also updated.
An error will be thrown if "cond_and_precond" is set, but also "cond"
or "precond". Otherwise regular conditions and preconditions will be
added as provided.

Skills still need to be adapted, especially those that have a transition
like "cond=bla, precond=true" which would now create 2 jump conditions
including a precondition with "true" statement -> not what we want.