fvcams: use YUV420_PLANAR as default in V4L2 with libv4l
If no specific format is requested and LIBV4L is available default to
the YU12 format which is the YUV420_PLANAR colorspace. This ensures that
we are compatible with the whole range of cameras supported by libv4l.
For example, cameras only providing a JPEG stream can now be used
properly, like the original Robotino Creative camera.