clips: add support to preload blackboard interface deftemplates
Add a (blackboard-preload ?type) function that can be used to create the
deftemplate for an interface before an interface is actually opened.
This makes it possible in some situtations to safe one level of file
loading. This requires to change the interface map data structure. We
need to also cover the case that the type has been asserted but not
interface yet, e.g. only a preload has happened. In the multimap we
cannot have that. Therefore we now have a map with list as second entry.