colli: SlowForward: Fix interpolation calculation of desired velocity
The previous version had a bug: It used the fixed value 0.7 for interpolation as y_0
and max_translation as y_1. Now if we set max_translation below 0.7, we get a totally
different interpolation behaviour (decreasing instead of increasing).

Now that we limit ourselves to M_PI/2 in order for driving forward at all, we can
simplyfy the interpolation. Also, it makes more sense to have a decreasing behaviour,
i.e. the faster we rotate, the lower the translation velocity should be.