colli: SlowForward: Calculate and adjust rotation/translation properly
The previous versions didn't adjust translation if the rotation was
cut due to set limits. It also checked the limits in a bad order, all
in all resulting in bad driving when the velocity-limits were set low
(e.g. driving in circle for relgoto(0,0.5)).

This is fixed now. It is also more readable, actually pretty simple
to understand.