fvutils: port VLC's colorthreshold
VLC has a video filter that detects colors which are "similar" to a
given reference color. It works with YUV422, thus requires no
conversion, and it's simple and fast. It has now been ported to fawkes
as a video filter. The actual similarity algorithm may need to be
factored out into a ColorModel to build efficient pipelines.