map-lasergen: allow laser data generation from current map pose
Up to now map-lasergen's purpose was to generate laser data to test the
amcl plugin. Now map-lasergen can also be used to create laser data
based on the actual current position (determined by amcl). This can, for
example, be useful for collision avoidance for static obstacles and a
limited sensor field of view.