aspects: deferred initialization of publisher in TransformAspect
Ther can be cases (as it is with map-lasergen) where you only know in
the init() method with access to the config whether the publisher is
required or not. To avoid having unused interfaces dangling around
support deferred initialization. In that case the aspect user indicates
that he might want to use the publisher. In that case he can then enable
the publishing part during the init() method.