Revert "colli: add bool for Drive(0,0) (just send once)"
This reverts commit fde6be22c61c2b6e431158d18c5fc7bca29ddf7c.

"just send once" is not the precise description of the commit, it would
rather be "just accept the received once". The problem is: Drive(..)
accelerates/decelerates towards the desired velocities *once*. It tries
to smoothly achieve those velocities. Hence if we e.g. want to stop, we
need to send as many Drive(0,0) until the robot stopped.
Now if the robot is currently faster than a "1-step-deceleration" would
make it stop, any following Drive(0,0) won't have any effect
=> there is no way for the robot to stop! Robot is on a rampage.

My intuition was wrong, this *is* unfortunately a very harmful commit, so
before I provide a proper fix for the actual problam that was intended
to be solved (see mailing-list for details), I had to undo this asap.