colli: improve behaviour in case colli needs to abort
If there is some problem (e.g. laser/motor is dead), the robot shuold
abort. Up to now, each abortion came with a "stop". We do not have that
now, we only want to stop if the colli was moving.

In addidion to that, not having a new target was considered as
"abortion", too. That caused the colli to send "stop" even after
being finished with the movement. Now the colli only stops the robot
if it needs to be stopped (e.g. error, target_reached,...) *and* the
colli is responsible for the current movement.