colli: remove "infamous" IsPipe() method
That method was used for caesar, to decide which laser readings
should be ignored because they were blocked by the "pipes" that
hold the construction of the robot.
We have that integrated in the laser plugin now, using deadspots.
I added a method to check if the reading is valid, i.e. if it is
higher than a low threshold (1cm). All invalid readings are filtered
by the laser plugin and have a distance of 0.