colli roboshape: fix and extend calculations of distances to robot edge
Besides fixes due to changes in coordinate-system compared to RCSoftX
(which had positive rotation clockwise), the calculation of the distance
from the laser to the edges of the robot has been simplified (less code,
hopefully easier to understand/read).

In addition, several methods have been added that can return angles to
the corners of the robot-shape (usefule for example in escape_drive_move).